refactor(chat): dedupe conversation queries for Sonar + mobile API
Extract shared creator conversation list/detail queries and route both
web and mobile chat handlers through the shared module.
