+++ apps/web/app/api/chat/conversations/route.ts
@@ -1,16 +1,10 @@
-import { count, desc, sql as drizzleSql, eq } from 'drizzle-orm';
+import { count, eq } from 'drizzle-orm';
 import { NextResponse } from 'next/server';
 import { getSessionContext } from '@/lib/auth/session';
-import {
-  sanitizeConversationTitle,
-  withSanitizedConversationTitles,
-} from '@/lib/chat/title';
+import { listCreatorConversations } from '@/lib/chat/conversation-queries';
+import { sanitizeConversationTitle } from '@/lib/chat/title';
 import { db } from '@/lib/db';
-import {
-  chatConversations,
-  chatMessages,
-  chatTurns,
-} from '@/lib/db/schema/chat';
+import { chatConversations, chatMessages } from '@/lib/db/schema/chat';
 import { captureError } from '@/lib/error-tracking';
 import { logger } from '@/lib/utils/logger';
 import { getSessionErrorResponse } from '../session-error-response';
@@ -50,29 +44,14 @@ export async function GET(req: Request) {
       ? Math.min(Math.max(parsed, 1), 50)
       : 20;
 
-    const conversations = await db
-      .select({
-        id: chatConversations.id,
-        title: chatConversations.title,
-        createdAt: chatConversations.createdAt,
-        updatedAt: chatConversations.updatedAt,
-        latestMessageRole:
-          drizzleSql<string>`(SELECT ${chatMessages.role} FROM ${chatMessages} WHERE ${chatMessages.conversationId} = ${chatConversations.id} ORDER BY ${chatMessages.createdAt} DESC LIMIT 1)`.as(
-            'latest_message_role'
-          ),
-        latestTurnStatus:
-          drizzleSql<string>`(SELECT ${chatTurns.status} FROM ${chatTurns} WHERE ${chatTurns.conversationId} = ${chatConversations.id} ORDER BY ${chatTurns.updatedAt} DESC LIMIT 1)`.as(
-            'latest_turn_status'
-          ),
-      })
-      .from(chatConversations)
-      .where(eq(chatConversations.creatorProfileId, profile.id))
-      .orderBy(desc(chatConversations.updatedAt))
-      .limit(limit);
+    const conversations = await listCreatorConversations({
+      creatorProfileId: profile.id,
+      limit,
+    });
 
     return NextResponse.json(
       {
-        conversations: withSanitizedConversationTitles(conversations),
+        conversations,
       },
       { status: 200, headers: NO_STORE_HEADERS }
     );
